Committing in .
 Patches for VMS

 Modified Files:
 	wxWindows/setup.h_vms wxWindows/include/wx/image.h
 	wxWindows/src/common/datetime.cpp
 	wxWindows/src/common/descrip.mms
 	wxWindows/src/generic/descrip.mms
 	wxWindows/src/unix/threadpsx.cpp
 ----------------------------------------------------------------------


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@5075 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/image.h
@@ -185,7 +185,7 @@ public:
   static void InitStandardHandlers();
 
   // GRG: Dic/99
-  unsigned long CountColours( unsigned long stopafter = -1 );
+  unsigned long CountColours( unsigned long stopafter = (unsigned long) -1 );
   unsigned long ComputeHistogram( wxHashTable &h );
